feat: separate logout logic from header component into custom hook
